refactor(fixed directory routes in client): fixed directory routes in client
