+++ app/client/components/searchPage/jobDetails.jsx
