+++ app/client/components/searchPage/jobList.jsx
