+++ app/client/containers/dashboard.jsx
@@ -6,10 +6,10 @@ import Container from 'muicss/lib/react/container';
 import Row from 'muicss/lib/react/row';
 import Col from 'muicss/lib/react/col';
 import Panel from 'muicss/lib/react/panel';
-import JobStepper from './jobStepper';
-import JobTable from './dashboard/jobTable';
-import { updateJobStatusAPI, deleteJobAPI, fetchUserJobs } from '../actions/actions';
-import '../../public/assets/css/dashboard.css';
+import JobStepper from '../shared/jobStepper/jobStepper';
+import JobTable from '../components/dashboard/jobTable';
+import { updateJobStatusAPI, deleteJobAPI, fetchUserJobs } from '../../actions/actions';
+import '../../../public/assets/css/dashboard.css';
 
 class Dashboard extends Component {
   constructor(props) {
