+++ app/client/root.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Provider } from 'react-redux';
-import App from './app';
+import App from './containers/app';
 import store from '../store/store';
 
 const Root = () => (
