+++ app/client/shared/jobStepper/jobStepper.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import Stepper from './dashboard/horizontalStepper/stepper';
+import Stepper from './horizontalStepper/stepper';
 
 const stepsTitles = [{ title: 'Start' }, { title: 'Application' }, { title: 'Submit' }, { title: 'Interview' }, { title: 'Offer' }];
