+++ app/index.jsx
@@ -2,7 +2,7 @@
 import React from 'react';
 import { render } from 'react-dom';
 import { BrowserRouter as Router } from 'react-router-dom';
-import App from './components/app';
+import App from './client/containers/app';
 
 
 const Index = () => (
